Fixed flow i.d overflow error

+++ cdh/ieu/fsw/src/tasks/cmd_apps/cmd_ers/cmd_ers.c
@@ -80,8 +80,8 @@ void cmd_ers(void* arg) {
     rt_printf("%d (CMD_ERS_TASK) Task started\n",time(NULL));
 
     // Declarations and initialization:
-    int8_t ret_val; // Function return value
-    int8_t flw_id;  // Flow identifier returned by rt_task_receive
+    int8_t   ret_val; // Function return value
+    uint32_t flw_id;  // Flow identifier for the message transaction
 
     uint16_t cmd_apid;           // Command APID
     uint16_t cmd_pkt_name;       // Command packet name
